Fix PHPStan errors in the data factories classes
[MAILPOET-4037]

+++ mailpoet/tests/DataFactories/WooCommerceProduct.php
@@ -165,8 +165,14 @@ class WooCommerceProduct {
 
   public function deleteAll() {
     $list = $this->tester->cliToArray(['wc', 'product', 'list', '--format=json', '--user=admin', '--fields=id']);
-    foreach (json_decode($list[0], true) as $item) {
-      $this->delete($item['id']);
+    $items = json_decode($list[0], true);
+
+    if (is_array($items)) {
+      foreach ($items as $item) {
+        if (is_array($item) && is_int($item['id'])) {
+          $this->delete($item['id']);
+        }
+      }
     }
   }
